uniforms uint surf user color in mesh render settings

+++ src/ext/qt/gui/mesh_render_settings_frame.cpp
@@ -215,7 +215,7 @@ void MeshRenderSettingsFrame::updatePointsTabFromSettings()
         ui->pointShadingCircleRadioButton->setEnabled(false);
 
         updatePointsColorComboBoxFromSettings();
-        ui->pointSizeSlider->setValue(mrs.pointWidth());
+        ui->pointSizeSlider->setValue((uint) mrs.pointWidth());
     }
     else {
         ui->pointsTab->setEnabled(false);
